+++ .travis.yml
@@ -18,7 +18,7 @@ before_install:
   - if [ "$TRAVIS_REPO_SLUG" = twbs-savage/bootstrap ]; then export TWBS_DO_VALIDATOR=0; fi
 install:
   - bundle install --deployment --jobs=3
-  - cp grunt/npm-shrinkwrap.json ./
+  - cp gulp/npm-shrinkwrap.json ./
   - npm install
 cache:
   directories:
