+++ package.json
@@ -17,14 +17,8 @@
     "Twitter, Inc."
   ],
   "scripts": {
-    "change-version": "node grunt/change-version.js",
-    "eslint": "eslint --config js/.eslintrc.json js/src",
-    "jscs": "jscs --config=js/.jscsrc js/src js/tests/unit docs/assets/js/src grunt Gruntfile.js docs/assets/js/ie-emulation-modes-warning.js docs/assets/js/ie10-viewport-bug-workaround.js",
-    "htmlhint": "htmlhint --config docs/.htmlhintrc _gh_pages/",
-    "postcss": "postcss --config grunt/postcss.js --replace dist/css/*.css",
-    "postcss-docs": "postcss --config grunt/postcss.js --no-map --replace docs/assets/css/docs.min.css && postcss --config grunt/postcss.js --no-map --replace docs/examples/**/*.css",
-    "shrinkwrap": "npm shrinkwrap --dev && shx mv ./npm-shrinkwrap.json ./grunt/npm-shrinkwrap.json",
-    "test": "npm run eslint && npm run jscs && grunt test"
+    "change-version": "node gulp/change-version.js",
+    "shrinkwrap": "npm shrinkwrap --dev && shx mv ./npm-shrinkwrap.json ./gulp/npm-shrinkwrap.json"
   },
   "style": "dist/css/bootstrap.css",
   "sass": "scss/bootstrap.scss",
@@ -42,38 +36,38 @@
     "tether": "^1.1.1"
   },
   "devDependencies": {
-    "autoprefixer": "^6.0.3",
     "babel-eslint": "^6.0.4",
+    "build-control": "^0.3.8",
     "eslint": "^3.0.0",
-    "grunt": "^1.0.1",
-    "grunt-babel": "^5.0.3",
-    "grunt-build-control": "^0.7.0",
-    "grunt-contrib-clean": "^1.0.0",
-    "grunt-contrib-compress": "^1.1.0",
-    "grunt-contrib-concat": "^1.0.0",
-    "grunt-contrib-connect": "^1.0.0",
-    "grunt-contrib-copy": "^1.0.0",
-    "grunt-contrib-cssmin": "^1.0.0",
-    "grunt-contrib-qunit": "^1.0.1",
-    "grunt-contrib-sass": "^1.0.0",
-    "grunt-contrib-uglify": "^1.0.0",
-    "grunt-contrib-watch": "^1.0.0",
-    "grunt-exec": "^1.0.0",
-    "grunt-html": "^8.0.1",
-    "grunt-jekyll": "^0.4.2",
-    "grunt-sass": "^1.0.0",
-    "grunt-saucelabs": "^9.0.0",
-    "grunt-scss-lint": "^0.3.8",
-    "grunt-stamp": "^0.3.0",
+    "gulp": "^3.9.1",
+    "gulp-babel": "^6.1.2",
+    "gulp-clean": "^0.3.2",
+    "gulp-clean-css": "^2.0.11",
+    "gulp-concat": "^2.6.0",
+    "gulp-connect": "^4.1.0",
+    "gulp-debug": "^2.1.2",
+    "gulp-eslint": "^3.0.1",
+    "gulp-exec": "^2.1.2",
+    "gulp-html": "^0.4.4",
+    "gulp-htmlhint": "^0.3.1",
+    "gulp-jscs": "^4.0.0",
+    "gulp-jshint": "^2.0.1",
+    "gulp-postcss": "^6.1.1",
+    "gulp-qunit": "^1.4.0",
+    "gulp-rename": "^1.2.2",
+    "gulp-ruby-sass": "^2.0.6",
+    "gulp-scss-lint": "^0.4.0",
+    "gulp-sourcemaps": "^1.6.0",
+    "gulp-stamp": "0.0.9",
+    "gulp-uglify": "^1.5.4",
     "htmlhint": "^0.9.13",
     "is-travis": "^1.0.0",
     "jscs": "^3.0.4",
-    "load-grunt-tasks": "^3.4.0",
-    "postcss-cli": "^2.5.2",
+    "jshint": "^2.9.2",
     "postcss-flexbugs-fixes": "^2.0.0",
     "shelljs": "^0.7.0",
     "shx": "^0.1.2",
-    "time-grunt": "^1.2.1"
+    "time-require": "^0.1.2"
   },
   "engines": {
     "node": ">=4"
